+++ clients/admin-ui/src/features/common/helpers.ts
@@ -92,7 +92,7 @@ export function getErrorFromResult(result: RTKReturnType) {
     if (isErrorWithDetail(error)) {
       errorMsg = error.data.detail;
     } else if (isErrorWithDetailArray(error)) {
-      errorMsg = error.data.detail[0].msg;
+      errorMsg = `${error.data.detail[0].msg}: ${error.data.detail[0].loc}`;
     } else if (isConflictError(error)) {
       errorMsg = `${error.data.detail.error} (${error.data.detail.fides_key})`;
     }