+++ clients/admin-ui/src/features/dataset/dataset.slice.ts
@@ -22,6 +22,11 @@ const initialState: State = {
   activeFieldIndex: null,
 };
 
+interface DatasetDeleteResponse {
+  message: string;
+  resource: Dataset;
+}
+
 export const datasetApi = createApi({
   reducerPath: "datasetApi",
   baseQuery: fetchBaseQuery({
@@ -49,11 +54,21 @@ export const datasetApi = createApi({
       }),
       invalidatesTags: ["Dataset"],
     }),
-    createDataset: build.mutation<Dataset, Dataset>({
+    // we accept 'unknown' as well since the user can paste anything in, and we rely
+    // on the backend to do the validation for us
+    createDataset: build.mutation<Dataset, Dataset | unknown>({
       query: (dataset) => ({
         url: `dataset/`,
         method: "POST",
         body: dataset,
+      }),
+      invalidatesTags: ["Datasets"],
+    }),
+    deleteDataset: build.mutation<DatasetDeleteResponse, FidesKey>({
+      query: (key) => ({
+        url: `dataset/${key}`,
+        params: { resource_type: "dataset" },
+        method: "DELETE",
       }),
       invalidatesTags: ["Datasets"],
     }),
@@ -65,6 +80,7 @@ export const {
   useGetDatasetByKeyQuery,
   useUpdateDatasetMutation,
   useCreateDatasetMutation,
+  useDeleteDatasetMutation,
 } = datasetApi;
 
 export const datasetSlice = createSlice({